Perform replication retries on a fixed interval
Implement a simple retry mechanism based on a thread per follower which
attempts to send out failed replication requests on a fixed interval.
The thread listens on an MPSC receiver, which is sent a message when a
replication request fails for that follower. Therefore the thread will
spend most of its time asleep unless a follower goes down or is
unreachable.