Use timestamps to handle out of order replica updates
There are cases where an older update for a particular key can arrive
at the replica after the newer update. For example, a successful
replicated SET operation can occur after the follower stops being
unreachable but before the next pending-updates retry attempt. In order
to correctly save the most up-to-date value, we associate each stored
entry with a "timestamp" to determine the correct order of updates. This
timestamp is really the number of milliseconds since some arbitrary
instant in time (in this case, when the leader is created). We use the
std::time::Instant API to ensure monotonicity of the timestamps, which
is necessary for the updates to be correctly ordered.